More pizo adjustments improvements
The progressive alarm increases volume gradually between volume levels
using linear interpolation.  The interpolation is now done with 16-bit
precision resulting in a smoother increase in volume at lower volumes.

+++ pizo.h
@@ -40,7 +40,7 @@
 
 typedef struct {
     uint8_t  status;
-    uint8_t  cm_factor;
+    uint16_t cm_factor;
     uint16_t timer;
 
     uint8_t  pos;
